% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{ea_icons}
\alias{ea_icons}
\alias{ea_icons_search}
\title{Icon}
\usage{
ea_icons(names)

ea_icons_search(names)
}
\arguments{
\item{names}{The name or vector of names of the icon(s) as listed in \link{icons}.}
}
\description{
Get an icon from the database of \link{icons}.
}
\examples{
ea_icon("bar_graph")

}
