% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\docType{data}
\name{icons}
\alias{icons}
\title{Icons.}
\format{A data frame with 77 rows and 2 variables:
\describe{
  \item{path}{SVG path of icons}
  \item{name}{Name of icon}
}}
\source{
\url{https://dribbble.com/shots/1934932-77-Essential-Icons-Free-Download}
}
\description{
This is a dataset 77 SVG icons.
}
\keyword{data}
