% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ea_asset}
\alias{ea_asset}
\alias{ea_convert}
\alias{ea_copy}
\alias{ea_source}
\title{Assets}
\usage{
ea_asset(asset, convert = TRUE)

ea_convert(file)

ea_copy(asset, dir = "www")

ea_source(asset, dir = "www")
}
\arguments{
\item{asset}{Name of texture, see details below.}

\item{convert}{Converts image to JSON formatted arrays.}

\item{file}{Path to file.}

\item{dir}{Directory to copy asset to.}
}
\description{
Assets for the \href{http://echarts4r.john-coene.com}{echarts4r} package.
}
\details{
Due to browser
"same origin policy" security restrictions, loading textures
from a file system may lead to a security exception,
see
\url{https://github.com/mrdoob/three.js/wiki/How-to-run-things-locally}.
References to file locations work in Shiny apps, but not in stand-alone
examples. The \code{*texture} functions facilitates transfer of image
texture data from R into textures when \code{convert} is set to \code{TRUE}.

Valid run \code{\link{ea_bank}} \code{asset}.
}
\section{Functions}{

\itemize{
  \item{\code{ea_asset} include asset.}
  \item{\code{ea_convert} convert file to JSON formatted array.}
  \item{\code{ea_copy} moves an asset to the folder of your choice. This is useful as \code{convert = TRUE} argment is computationally expensive, for instance, in Shiny, move the asset to your www folder for better perfromances.}
  \item{\code{ea_source} source your copied asset.}
}
}

\examples{
\dontrun{
library(shiny)
library(echarts4r)

# copy asset to www folder
ea_copy("world")

ui <- fluidPage(
  echarts4rOutput("globe")
)

server <- function(input, output){

  output$globe <- renderEcharts4r({

    e_charts() \%>\%
      e_globe(
        environment = ea_source("earth"),
        base.texture = e_globe_texture()
      )
  })

}

shinyApp(ui, server)
}

}
